diarization-v2: Updating scripts

+++ egs/wsj/s5/steps/diarization/convert_labels_to_data.sh
@@ -9,12 +9,11 @@ max_segment_length=1000
 overlap_length=100
 
 data=$1
-ivector_dir=$2
-cluster_dir=$3
-out_data=$4
+cluster_dir=$2
+out_data=$3
 
-if [ $# -ne 4 ]; then
-  echo "Usage: $0 <data> <ivector-dir> <cluster-dir> <out-data>"
+if [ $# -ne 3 ]; then
+  echo "Usage: $0 <data> <cluster-dir> <out-data>"
   exit 1
 fi
 
@@ -22,17 +21,17 @@ rm -r $out_data || true
 mkdir -p $out_data
 
 for f in wav.scp reco2file_and_channel glm stm; do
-  cp $data/$f $out_data
+  [ -f $f ] && cp $data/$f $out_data
 done
 
-utils/data/get_reco2utt.sh $ivector_dir
+utils/data/get_reco2utt.sh $data
 
 $cmd $cluster_dir/log/get_segments.log \
   segmentation-init-from-segments --shift-to-zero=false --frame-overlap=0.0 \
   --utt2label-rspecifier=ark,t:$cluster_dir/labels \
-  $ivector_dir/segments ark:- \| \
+  $data/segments ark:- \| \
   segmentation-combine-segments-to-recordings ark:- \
-  ark,t:$ivector_dir/reco2utt \
+  ark,t:$data/reco2utt \
   ark:- \| \
   segmentation-post-process --merge-adjacent-segments ark:- ark:- \| \
   segmentation-split-segments \
